+++ jtag/src/writemem.c
@@ -37,7 +37,7 @@ void
 writemem( bus_t *bus, FILE *f, uint32_t addr, uint32_t len )
 {
 	uint32_t step;
-	uint64_t a;
+	uint32_t a;
 	int bc = 0;
 	int bidx = 0;
 #define BSIZE 4096
